requirements.txt
google-generativeai
langchain
PyPDF2
faiss-cpu
streamlit
langchain-google-genai
dotenv